requirements.txt
fastapi
google-generativeai
pandas
uvicorn[standard]
pydantic